fix(connectivity): address Wi-Fi recovery review feedback
