+++ firmware/host/app/__tests__/boot-network-recovery.test.ts
@@ -19,6 +19,11 @@ test('boot Wi-Fi helper detects scan exhaustion from network-service failures', 
   assert.equal(bootWiFiFailureMessage('Access point "home-ap" not found'), '保存済みWi-Fiが見つかりません')
 })
 
+test('boot Wi-Fi helper classifies non-scan failures as connection-failed', () => {
+  assert.equal(classifyBootWiFiFailure('authentication failed'), 'connection-failed')
+  assert.equal(bootWiFiFailureMessage('authentication failed'), 'Wi-Fi接続に失敗しました')
+})
+
 test('boot Wi-Fi helper maps offline choice to a skipped network readiness result', () => {
   assert.deepEqual(networkReadyResultForRecoveryChoice('offline', 'authentication failed'), {
     status: 'skipped',
