+++ firmware/host/modules/testing/fakes/capabilities.ts
@@ -12,3 +12,16 @@ export type RobotLed = {
   blink(r: number, g: number, b: number, duration: number, index?: number, count?: number): void
   rainbow(index?: number, count?: number): void
 }
+
+export type NetworkReadyResult =
+  | {
+      status: 'connected'
+    }
+  | {
+      status: 'skipped'
+      reason: string
+    }
+  | {
+      status: 'failed'
+      reason: string
+    }
